+++ src/Spinner/stories.js
@@ -1,0 +1,18 @@
+/* eslint-disable react/prop-types */
+import React from 'react';
+
+import { Spinner } from './index';
+
+export default { title: 'Spinner', component: Spinner };
+
+export const Default = () => <Spinner />;
+export const Large = () => <Spinner large />;
+export const Light = () => (
+  <div
+    style={{
+      backgroundColor: '#1e1e1e',
+    }}
+  >
+    <Spinner light />
+  </div>
+);